move gpus to separate files to avoid duplicate symbols when compiling as part of legacy mangohud
